Add .gitignore files and bg_map scene
The commit introduces two .gitignore files containing patterns of files and directories that Git should ignore. Additionally, a new scene 'bg_map' is added which includes a Node2D with a TileMap child node and related configurations. The structure of this scene seems suggestive of a background setting for the project.