Add enemy and bullet classes with movement functionality
This commit includes the creation of the Bullet and Enemy classes, with movement functionality being added to both. The Bullet class uses direction and speed to determine its movement, whereas the Enemy class uses the Player's position - obtained through Godot's GetFirstNodeInGroup - for directionality. In addition, sprites and animations for the Enemy class were added. Lastly, bullet shooting functionality was also added to the Weapon class.