Add comprehensive enemy actions and effects
A new shader has been added to enable visual effects when an enemy is damaged. The 'Enemy.cs' script was updated with an 'EnemyHurt' function that determines the next actions when an enemy is hurt, including playing a hurt animation and reducing the enemy's HP. The 'Bullet.cs' script was modified to call the new 'EnemyHurt' function. Also, the 'Animations.cs' script was updated with a 'RunAnimation' function, used for running various animations.